+++ test/test_ad_licenselint.rb
@@ -2,8 +2,11 @@ $LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
 require 'minitest/autorun'
 require 'minitest/hooks/test'
 require 'ad_licenselint'
+require 'yaml'
+require 'fileutils'
 
 PODFILE_DIR = "test/TestLicenseCollector"
+CONFIG_PATH = File.join(PODFILE_DIR, ".ad_licenselint.yml")
 
 class TestCase < Minitest::Test
   include Minitest::Hooks
@@ -14,6 +17,10 @@ class ADLincenseLintTest < TestCase
   # from minitest-hooks
   def after_all
     install_pods(["Alamofire"])
+  end
+
+  def teardown
+    FileUtils.rm CONFIG_PATH if File.exist? CONFIG_PATH
   end
 
   def test_one_valid_pod
@@ -54,6 +61,28 @@ class ADLincenseLintTest < TestCase
       content = ADLicenseLint::Runner.new(options).run
       assert_includes content, "ObjectivePGP"
       assert_includes content, "BSD for non-commercial use"
+    }
+  end
+
+  def test_one_invalid_pod_with_allowlist
+    install_pods(["ObjectivePGP"])
+    write_allowlist(["ObjectivePGP", "OtherPod"])
+
+    [false, true].each { |all|
+      options = {
+        format: ADLicenseLint::Constant::TERMINAL_FORMAT_OPTION,
+        path: PODFILE_DIR,
+        all: all
+      }
+
+      content = ADLicenseLint::Runner.new(options).run
+      if all
+        assert_includes content, "ObjectivePGP"
+        assert_includes content, "BSD for non-commercial use"
+      else
+        refute_includes content, "ObjectivePGP"
+        refute_includes content, "BSD for non-commercial use"
+      end
     }
   end
 
@@ -139,5 +168,10 @@ end"""
     Dir.chdir(PODFILE_DIR) do
       system "bundle exec pod install > /dev/null"
     end
+  end
+
+  def write_allowlist(pods)
+    config = { "allow" => pods }
+    File.write(CONFIG_PATH, config.to_yaml)
   end
 end